% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{delete_old_files}
\alias{delete_old_files}
\title{Cleanup old files}
\usage{
delete_old_files(before = "2021-12-03")
}
\arguments{
\item{before}{date before for which to delete older files}
}
\description{
Remove old files from the CRANlike server.
}
