% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuilds.R
\name{trigger_rebuilds}
\alias{trigger_rebuilds}
\title{Rebuild packages}
\usage{
trigger_rebuilds(
  repository = "r-universe/r-lib",
  workflow = "build.yml",
  days = 30
)
}
\arguments{
\item{repository}{name of the github repository}

\item{workflow}{name of the workflow to trigger}

\item{days}{trigger rebuild every n days}
}
\description{
Automatically trigger workflows once every n days.
}
