% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuilds.R
\name{trigger_rebuilds}
\alias{trigger_rebuilds}
\alias{rebuild_vignettes}
\alias{rebuild_oldies}
\alias{rebuild_missing_binaries}
\alias{cancel_queued_builds}
\alias{cancel_all_queued_builds}
\alias{delete_one}
\alias{rebuild_all_remotes}
\alias{rebuild_universe_remotes_only}
\title{Rebuild packages}
\usage{
trigger_rebuilds(
  repository = "r-universe/jeroen",
  workflow = "build.yml",
  days = 30
)

rebuild_vignettes(universe = "jeroen")

rebuild_oldies(universe, before = "2021-03-14")

rebuild_missing_binaries(universe = "ropensci")

cancel_queued_builds(universe = "ropensci")

cancel_all_queued_builds()

delete_one(universe, pkg)

rebuild_all_remotes()

rebuild_universe_remotes_only(universe)
}
\arguments{
\item{repository}{name of the github repository}

\item{workflow}{name of the workflow to trigger}

\item{days}{trigger rebuild every n days}

\item{universe}{name of the universe, use NULL for all universes}

\item{before}{date before which to rebuild}

\item{pkg}{name of package to delete}
}
\description{
Automatically trigger workflows once every n days.
}
