% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuilds.R
\name{trigger_rebuilds}
\alias{trigger_rebuilds}
\alias{rebuild_vignettes}
\alias{rebuild_oldies}
\title{Rebuild packages}
\usage{
trigger_rebuilds(
  repository = "r-universe/jeroen",
  workflow = "build.yml",
  days = 30
)

rebuild_vignettes(universe = "jeroen")

rebuild_oldies(universe, before = "2021-03-01")
}
\arguments{
\item{repository}{name of the github repository}

\item{workflow}{name of the workflow to trigger}

\item{days}{trigger rebuild every n days}

\item{universe}{name of the universe, use NULL for all universes}

\item{before}{date before which to rebuild}
}
\description{
Automatically trigger workflows once every n days.
}
